+++ src/ios/DogView/DogView.css.js
@@ -1,0 +1,13 @@
+//@flow
+'use strict';
+import React, {
+  StyleSheet,
+} from 'react-native';
+
+const DogViewCSS = StyleSheet.create({
+  dogView: {
+    flex: 1,
+  },
+});
+
+module.exports = DogViewCSS;
