+++ src/ios/LeaderButton/LeaderButton.css.js
@@ -1,0 +1,18 @@
+//@flow
+'use strict';
+import React, {
+  StyleSheet,
+} from 'react-native';
+
+const LeaderButtonCSS = StyleSheet.create({
+  leaderBoardRibbon: {
+    height: 35,
+    width: 35,
+  },
+
+  leaderBoardView: {
+    margin: 15,
+  }
+});
+
+module.exports = LeaderButtonCSS;
