+++ src/ios/LeaderButton/ribbon.png
